requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
cffi==1.15.0
chardet==4.0.0
cycler==0.11.0
Django==4.0
fonttools==4.28.3
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
mpmath==1.2.1
multidict==5.2.0
numpy==1.21.4
packaging==21.3
Pillow==8.4.0
py-cord @ git+https://github.com/Pycord-Development/pycord@c50d0b9c73b98c766998ce5a5574442e3afa2fe8
pycparser==2.21
PyNaCl==1.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
six==1.16.0
sqlparse==0.4.2
sympy==1.9
typing-extensions==4.0.1
tzdata==2021.5
yarl==1.7.2
